Avance del proyecto hasta contratos.
Agrega contratos con un cliente ingresando el pdf del contrato, agrega
retencion y lo edita, edita el contrato y lo elimina.

+++ app/database/migrations/2016_09_14_124751_crear_tabla_contratos.php
@@ -23,7 +23,7 @@ class CrearTablaContratos extends Migration {
 				->onUpdate('cascade')->onDelete('cascade');
 			$table->date('inicio');
 			$table->date('fin');
-			$table->float('total');
+			$table->double('total', 15, 2);
 			$table->boolean('igv')->nullable();
 		});
 	}
